# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dexe Investing Api",
        "title": "Dexe Investing Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/user/signup": {
            "post": {
                "description": "User SignUp [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "User SignUp [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User nickname",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet",
                        "name": "wallet",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with image",
                        "name": "file",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}": {
            "get": {
                "description": "Get User Info",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}/avatar": {
            "put": {
                "description": "Update User Avatar [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Avatar [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Body with image",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        },
        "/user/{wallet}/nickname": {
            "put": {
                "description": "Update User Nickname [SIGN]",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update User Nickname [SIGN]",
                "parameters": [
                    {
                        "type": "string",
                        "default": "iQxX3slnRg",
                        "description": "An authorization header",
                        "name": "x-morph",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "User wallet address",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "New user nickname",
                        "name": "nickname",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.S"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.E"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "wallet": {
                    "type": "string"
                }
            }
        },
        "response.E": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "context": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.S": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object"
                },
                "datetime": {
                    "type": "string"
                }
            }
        },
...
    }
}